# Benchmarking performance impact of capturing identifiers during parsing