--- a/src/Compiler/Driver/ParseAndCheckInputs.fs
+++ b/src/Compiler/Driver/ParseAndCheckInputs.fs
@@ -413,7 +413,8 @@ let ParseInput
         lexbuf: UnicodeLexing.Lexbuf,
         defaultNamespace,
         fileName,
-        isLastCompiland
+        isLastCompiland,
+        identCapture
     ) =
     // The assert below is almost ok, but it fires in two cases:
     //  - fsi.exe sometimes passes "stdin" as a dummy file name
@@ -433,25 +434,29 @@ let ParseInput
         let input =
             let identStore = HashSet<string>()
 
-            let identCaptureLexer x =
-                let token = lexer x
+            let lexer =
+                if identCapture then
+                    (fun x ->
+                        let token = lexer x
 
-                match token with
-                | Parser.token.PERCENT_OP ident
-                | Parser.token.FUNKY_OPERATOR_NAME ident
-                | Parser.token.ADJACENT_PREFIX_OP ident
-                | Parser.token.PLUS_MINUS_OP ident
-                | Parser.token.INFIX_AMP_OP ident
-                | Parser.token.INFIX_STAR_DIV_MOD_OP ident
-                | Parser.token.PREFIX_OP ident
-                | Parser.token.INFIX_BAR_OP ident
-                | Parser.token.INFIX_AT_HAT_OP ident
-                | Parser.token.INFIX_COMPARE_OP ident
-                | Parser.token.INFIX_STAR_STAR_OP ident
-                | Parser.token.IDENT ident -> identStore.Add ident |> ignore
-                | _ -> ()
+                        match token with
+                        | Parser.token.PERCENT_OP ident
+                        | Parser.token.FUNKY_OPERATOR_NAME ident
+                        | Parser.token.ADJACENT_PREFIX_OP ident
+                        | Parser.token.PLUS_MINUS_OP ident
+                        | Parser.token.INFIX_AMP_OP ident
+                        | Parser.token.INFIX_STAR_DIV_MOD_OP ident
+                        | Parser.token.PREFIX_OP ident
+                        | Parser.token.INFIX_BAR_OP ident
+                        | Parser.token.INFIX_AT_HAT_OP ident
+                        | Parser.token.INFIX_COMPARE_OP ident
+                        | Parser.token.INFIX_STAR_STAR_OP ident
+                        | Parser.token.IDENT ident -> identStore.Add ident |> ignore
+                        | _ -> ()
 
-                token
+                        token)
+                else
+                    lexer
 
             if FSharpMLCompatFileSuffixes |> List.exists (FileSystemUtils.checkSuffix fileName) then
                 if lexbuf.SupportsFeature LanguageFeature.MLCompatRevisions then
@@ -461,14 +466,14 @@ let ParseInput
 
             // Call the appropriate parser - for signature files or implementation files
             if FSharpImplFileSuffixes |> List.exists (FileSystemUtils.checkSuffix fileName) then
-                let impl = Parser.implementationFile identCaptureLexer lexbuf
+                let impl = Parser.implementationFile lexer lexbuf
 
                 let tripleSlashComments =
                     LexbufLocalXmlDocStore.ReportInvalidXmlDocPositions(lexbuf)
 
                 PostParseModuleImpls(defaultNamespace, fileName, isLastCompiland, impl, lexbuf, tripleSlashComments, Set identStore)
             elif FSharpSigFileSuffixes |> List.exists (FileSystemUtils.checkSuffix fileName) then
-                let intfs = Parser.signatureFile identCaptureLexer lexbuf
+                let intfs = Parser.signatureFile lexer lexbuf
 
                 let tripleSlashComments =
                     LexbufLocalXmlDocStore.ReportInvalidXmlDocPositions(lexbuf)
@@ -640,7 +645,8 @@ let ParseOneInputLexbuf (tcConfig: TcConfig, lexResourceManager, lexbuf, fileNam
                         lexbuf,
                         None,
                         fileName,
-                        isLastCompiland
+                        isLastCompiland,
+                        tcConfig.captureIdentifiersWhenParsing
                     )
 
                 // Report the statistics for testing purposes