--- a/src/Compiler/Service/IncrementalBuild.fsi
+++ b/src/Compiler/Service/IncrementalBuild.fsi
@@ -266,7 +266,7 @@ type internal IncrementalBuilder =
         enableParallelCheckingWithSignatureFiles: bool *
         dependencyProvider: DependencyProvider option *
         parallelReferenceResolution: ParallelReferenceResolution *
-        captureIdentifiersWhenParsing: bool->
+        captureIdentifiersWhenParsing: bool ->
             NodeCode<IncrementalBuilder option * FSharpDiagnostic[]>
 
 /// Generalized Incremental Builder. This is exposed only for unit testing purposes.