--- a/tests/FSharp.Compiler.ComponentTests/FSharpChecker/FindReferences.fs
+++ b/tests/FSharp.Compiler.ComponentTests/FSharpChecker/FindReferences.fs
@@ -96,8 +96,12 @@ secondA.DoNothing(secondB)
 
         }
 
-[<Fact>]
-let ``Finding references in project`` () =
+[<Theory>]
+[<InlineData(true, true)>]
+[<InlineData(true, false)>]
+[<InlineData(false, true)>]
+[<InlineData(false, false)>]
+let ``Finding references in project`` (fastCheck, captureIdentifiersWhenParsing) =
     let size = 20
 
     let project =
@@ -111,10 +115,12 @@ let ``Finding references in project`` () =
         |> updateFile "File005" (addDependency "File000")
         |> updateFile "File010" (addDependency "File000")
 
-    let checker = FSharpChecker.Create(enableBackgroundItemKeyStoreAndSemanticClassification = true)
+    let checker = FSharpChecker.Create(
+        enableBackgroundItemKeyStoreAndSemanticClassification = true,
+        captureIdentifiersWhenParsing = captureIdentifiersWhenParsing)
 
     project.WorkflowWith checker {
-        findAllReferencesToModuleFromFile "File000" true (expectNumberOfResults 5)
+        findAllReferencesToModuleFromFile "File000" fastCheck (expectNumberOfResults 5)
     }
 
 [<Fact>]