These are the results on my machine:

``` ini

BenchmarkDotNet=v0.13.2, OS=Windows 10 (10.0.19045.2311)
AMD Ryzen 9 5900X, 1 CPU, 24 logical and 12 physical cores
.NET SDK=7.0.100
  [Host]     : .NET 7.0.0 (7.0.22.51805), X64 RyuJIT AVX2 DEBUG
  Job-GIHPTA : .NET 7.0.0 (7.0.22.51805), X64 RyuJIT AVX2

InvocationCount=1  UnrollFactor=1  

```
|       Method | IdentCapture |     Mean |   Error |  StdDev | Allocated |
|------------- |------------- |---------:|--------:|--------:|----------:|
| **ParseBigFile** |        **False** | **105.0 ms** | **1.23 ms** | **1.15 ms** |   **30.2 MB** |
| **ParseBigFile** |         **True** | **107.3 ms** | **1.01 ms** | **0.90 ms** |   **32.8 MB** |

Seems it has a measurable impact. But just barely. Is it worth adding a flag to skip it for those who don't need it?