@auduchinok 2022-12-16
@0101 We do lots of additional parsing in Rider, and it would be great to not do extra work that could add additional GC pressure. We also don't plan to use this flag, so it seems it makes sense to make the parser also use it. Thank you for passing the flag to the parser!